almost there

+++ mand/help/usleep.c
@@ -1,17 +1,28 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   usleep.c                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: tamehri <tamehri@student.42.fr>            +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/04/04 17:51:42 by tamehri           #+#    #+#             */
+/*   Updated: 2024/04/04 18:20:15 by tamehri          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
 
 #include "../srcs/philos.h"
 
-size_t    get_current_time()
+long	get_current_time()
 {
-	struct timeval  tv;
+	struct timeval	tv;
 
 	gettimeofday(&tv, NULL);
 	return (tv.tv_sec * 1000 + tv.tv_usec / 1000);
 }
 
-void    ft_usleep(size_t micro)
+void	ft_usleep(long micro)
 {
-	size_t    start_time;
+	long	start_time;
 
 	start_time = get_current_time();
 	while (1)
